requirements.txt
django 
django-encrypted-model-fields
psycopg
djangorestframework
grpcio
daphne
Twisted[tls,http2]
adrf
channels
channels-redis
channels-redis[cryptography]
pillow
dramatiq
django_dramatiq
django-webpack-loader
pywebpush
cffi
protobuf
djangorestframework-simplejwt[crypto]
authlib
django-user-agents
pyyaml 
ua-parser 
user-agents
django-cors-headers
gunicorn
uvicorn